fixing provisioning of arm64 cloudimgs
