+++ aws/scenarios/microVMs/microvms/resources/arm64/domain.xls
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <xsl:stylesheet version="1.0"
                 xmlns:xsl="http://www.w3.org/1999/XSL/Transform">
-  <xsl:template match="@firmware" />
   <xsl:output omit-xml-declaration="yes" indent="yes"/>
+  <xsl:template match="@firmware" />
 
   <xsl:template match="node()|@*">
       <xsl:copy>
