+++ aws/scenarios/microVMs/microvms/resources/distro/domain.xml
@@ -1,0 +1,57 @@
+<?xml version="1.0"?>
+<xsl:stylesheet version="1.0"
+                xmlns:xsl="http://www.w3.org/1999/XSL/Transform">
+  <xsl:output omit-xml-declaration="yes" indent="yes"/>
+  <xsl:template match="node()|@*">
+      <xsl:copy>
+         <xsl:apply-templates select="node()|@*"/>
+      </xsl:copy>
+   </xsl:template>
+
+  <xsl:template match="/domain/features/pae">
+      <xsl:copy>
+          <xsl:apply-templates select="@* | node()"/>
+      </xsl:copy>
+      <gic version='3'/>
+  </xsl:template>
+
+
+   <xsl:template match="/domain/name"><name>{domainName}</name></xsl:template>
+  <xsl:template match="/domain/features">
+       <cpu mode='host-model' check='none'>
+           <model fallback='forbid'>qemu64</model>
+       </cpu>
+  </xsl:template>
+
+ <xsl:template match="/domain/os/type[@arch='aarch64']/@machine">
+      <xsl:attribute name="machine">
+          <xsl:value-of select="'virt'"/>
+      </xsl:attribute>
+  </xsl:template>
+
+  <xsl:template match="/domain/devices/disk">
+      <filesystem type='mount' accessmode='passthrough'>
+          <source dir='{sharedFSMount}'/>
+          <target dir='kernel-version-testing'/>
+      </filesystem>
+      <readonly/>
+      <channel type='unix'>
+          <source mode='bind' path='/tmp/{domainID}.sock'/>
+          <target type='virtio' name='logs-output-socket' state='disconnected'/>
+      </channel>
+      <xsl:copy>
+          <xsl:apply-templates select="@*|node()"/>
+      </xsl:copy>
+  </xsl:template>
+
+  <xsl:template match="/domain/devices/interface[@type='network']/mac/@address">
+      <xsl:attribute name="address">
+          <xsl:value-of select="'{mac}'"/>
+      </xsl:attribute>
+  </xsl:template>
+
+  <xsl:template match="domain/devices/graphics"/>
+  <xsl:template match="domain/devices/audio"/>
+  <xsl:template match="domain/devices/video"/>
+</xsl:stylesheet>
+
